moved domian models from Areas, update namespaces
