+++ Programming-Tournament/Areas/Admin/Pages/Users/Create.cshtml
@@ -2,7 +2,7 @@
 @model CreateModel
 @using Programming_Tournament.Areas.Admin.Pages.Users
 @using Programming_Tournament.Areas.Admin.Models
-@using Programming_Tournament.Areas.Identity.Models
+@using Programming_Tournament.Models.Domain.User
 @{
 }
