+++ Programming-Tournament/Areas/Admin/Pages/Users/Create.cshtml.cs
@@ -10,6 +10,7 @@ using Programming_Tournament.Areas.Admin.Models;
 using Programming_Tournament.Areas.Identity.Managers;
 using Programming_Tournament.Areas.Identity.Models;
 using Programming_Tournament.Data;
+using Programming_Tournament.Models.Domain.User;
 
 namespace Programming_Tournament.Areas.Admin.Pages.Users
 {
