+++ Programming-Tournament/Areas/Admin/Pages/Users/Edit.cshtml
@@ -1,7 +1,7 @@
 ﻿@page "{id}"
 @model EditModel
 @using Programming_Tournament.Areas.Admin.Pages.Users
-@using Programming_Tournament.Areas.Identity.Models
+@using Programming_Tournament.Models.Domain.User
 @{
 }
