+++ Programming-Tournament/Areas/Admin/Pages/Users/Index.cshtml.cs
@@ -10,6 +10,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Programming_Tournament.Areas.Identity.Models;
 using Programming_Tournament.Data;
+using Programming_Tournament.Models.Domain.User;
 
 namespace Programming_Tournament.Areas.Admin.Pages.Users
 {
