+++ Programming-Tournament/Areas/Identity/Managers/ApplicationsManager.cs
@@ -1,5 +1,6 @@
 ﻿using Programming_Tournament.Areas.Identity.Models;
 using Programming_Tournament.Data;
+using Programming_Tournament.Models.Domain.User;
 using System;
 using System.Collections.Generic;
 using System.Linq;
