+++ Programming-Tournament/Areas/Identity/Managers/UsersManager.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Programming_Tournament.Areas.Admin.Models;
 using Programming_Tournament.Areas.Identity.Models;
 using Programming_Tournament.Data;
+using Programming_Tournament.Models.Domain.User;
 using System;
 using System.Collections.Generic;
 using System.Linq;
