+++ Programming-Tournament/Areas/Identity/Models/ApplicationUserDetailsPageModel.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Programming_Tournament.Models.Domain.User;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
