+++ Programming-Tournament/Areas/Identity/Pages/Application/Index.cshtml
@@ -1,6 +1,6 @@
 ﻿@page
 @model IndexModel
-@using Programming_Tournament.Areas.Identity.Models
+@using Programming_Tournament.Models.Domain.User
 @using Programming_Tournament.Areas.Identity.Pages.Application
 @{
     ViewData["Title"] = "Student Application";
