+++ Programming-Tournament/Models/Domain/User/ApplicationUser.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Programming_Tournament.Areas.Identity.Models
+namespace Programming_Tournament.Models.Domain.User
 {
     public class ApplicationUser : IdentityUser
     {
