+++ Programming-Tournament/Models/Domain/User/Faculty.cs
@@ -4,13 +4,21 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Programming_Tournament.Areas.Identity.Models
+namespace Programming_Tournament.Models.Domain.User
 {
-    public class Lectern
+    public class Faculty
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public long LecternId { get; set; }
+        public long FacultyId { get; set; }
 
         public string Name { get; set; }
+    }
+
+    public enum DegreeType
+    {
+        Bachelor,
+        Master,
+        PhD,
+        Unknown
     }
 }
