+++ Programming-Tournament/Models/Domain/User/Lectern.cs
@@ -4,12 +4,12 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Programming_Tournament.Areas.Identity.Models
+namespace Programming_Tournament.Models.Domain.User
 {
-    public class Curriculum
+    public class Lectern
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int CurriculumId { get; set; }
+        public long LecternId { get; set; }
 
         public string Name { get; set; }
     }
