+++ Programming-Tournament/Startup.cs
@@ -14,6 +14,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Programming_Tournament.Areas.Identity.Managers;
 using Programming_Tournament.Areas.Identity.Models;
+using Programming_Tournament.Models.Domain.User;
 
 namespace Programming_Tournament
 {
