+++ Programming-Tournament/Views/Shared/_LoginPartial.cshtml
@@ -1,5 +1,5 @@
 @using Microsoft.AspNetCore.Identity
-@using Programming_Tournament.Areas.Identity.Models
+@using Programming_Tournament.Models.Domain.User
 
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
